Initial database write capability.
Basically a proof of concept to see if we can write to the database.

The start of the functionaly to write the queue of files to the database
so that they can be processed. We want to write the entire file list to
database first and then process it. Processing will take place in
batches of 500. Load the first 500 entries, process each one, and then
remove those 500 entires. Repeat.

One major problem with getting a CSContext in a Quartz job is that there
is no User account associated since Quartz was started on servlet
initiation and not when a user visits a page. So I had to create an
empty user object, set it to be an SYS_ADMIN and use that to get a
CSContext. Not sure if it'll result in errors later on, but it works for
now.