Index generation now works.
We only store files that do NOT have a copyright tag. So if there is an
entry in the database for that file, then it doesn't have a copyright
tag.

Rename username field in ubc_ctlt_ca_files table to userid as user id
was easier to obtain than username.

We will not index users who don't have write access and who are not
course instructors or course ISS staff. Note that the UBC_Instructor and
UBC_ISS roles are UBC specific roles. I've allowed UBC_Instructor to
fall back to the normal Instructor role in order to make sure it still
still works in my dev vm which doesn't have the UBC roles.

The roles should be made configurable later.

File queue generation is now restricted only to the course content
collection directories.

Note that the index generated is cleared everytime we run the indexing
job so we start fresh. It might be better to update instead of starting
from scratch but don't have time for that.