Restrict index generation to a single node.
Really bad giant commit because I forgot to commit frequently.

There can be multiple servers running on production and we only want the
indexing job to run on one of them. This commit mostly implements that.
Unfortunately, a lot of other smaller enhancements got bundled in too.

One of the major changes is that last run statistics are now stored in
the ubc_ctlt_ca_hosts table. I did this after realizing that the
configuration file might not be synced across nodes. And so storing it
in the database is the only good way of communicating with other nodes.

Unfortunately, this brings up the question of *how* configuration can be
preserved across nodes if the stored configuration files aren't
propagated?! This will have to be resolved when I send the request to
run this.

There's also some early experimentation on creating a user facing
interface as an alerts module.