Convert Date datatype to Timestamp, setLeader fix
Apparently, the Date datatype only has resolution to 1 day and I need
resolution to 1 second. So had to change it to Timestamp.

Also added stack trace printing to exceptions, so they don't fail
silently.

Also changed setLeader in HostsTable to update the entry instead of
deleting and then recreating it. Deleting and recreating the entry
resulted in data loss as the prior run records was also deleted.