Add save indicators to system config UI.
Previously there was no indication whether clicking the save
configuration button actually did anything. Added some callbacks so that
the user would have an visual indication that something is happening.

The user will know when a save is in progress and when a save has
succeeded or failed. This save message will also remove itself after a
few seconds.