Move Copyright Alerts Module to jsp from servlet.
Cause I couldn't get the servlet to be used by the module page loader.
So now move it into a jsp. Should be fine if I keep it mostly js.

A lot of trouble had to be gone through to let us use AngularJS on a
module page as there's some sort of javascript loader being used by
Blackboard. It changes the load order and so messes up AngularJS
initialization. Switched to a custom solution that let's us control the
loading order and AngularJS initialization.