DB src reorganization, module view working.
Database related java source files was moved out of the systemconfig
package and into it's own parent package since both systemconfig and
alertsmodule now uses them.

Also can now display the copyright alert lists to the user in a view.

The view itself is pretty straightforward AngularJS.

Loading AngularJS on the other hand is complicated due to the javascript
loader used by blackboard on module pages. It strips all <script> tags
present in the module and executes them later once the page has been
loaded. I built a sorta intermediate loader that'll guarantee that
AngularJS and other components needed for the module is loaded in the
correct order so that they'll work.

Another problem was that I couldn't generate a url to the building
block servlets automatically, so just hardcoded them.