IE compatibility CSS fixes.
IE9 & 10 doesn't show the Copyright Alerts box even if there are files
to be tagged because it starts Angular twice for some reason. So the
hideInitially class gets removed and then added back again. Fixed this
by using removeClassName() instead of toggleClassName().

IE8 will only apply the styles if it comes after the content html. So I
had to move all the style declarations to after the content div.