Skip null CSEntry files.
Apparently, some specially named files, such as "xid-9052367_1.jpeg"
can't be accessed either through the API or through the Blackboard UI
for some reason. So they return "null" when you search for them even
though they show up in listings. Modified indexing to skip these files
otherwise it'd just crash with a NullPointerException.

When tracking that problem down, I added some more logging and decided
to keep them in.

Up the LOADNUM to 500, so we'll process and load 500 files at a time in
a batch for indexing.