Fix indexing not reading metadata attributes.
The indexing is supposed to skip files that had already been tagged but
it wasn't. It looks like it wasn't using the metadata attribute IDs
configured by the user. This turned out to be cause by the
SavedConfiguration object not loading saved settings when it was
created, so there were no attributes to be read. This had also broken
the runtime limit. The fix was to add a call to load after the object
was created.

There's a reason why I don't load the configuration in the
SavedConfiguration object's constructor, I think it was due to the
exceptions, but can't exactly remember. It might be good to revisit this
later on.