New Alerts Module API, allow file list pagination
Retrieving a few thousand files turns out to freeze my computer. Just in
case some users might have a few thousand files for whatever reason,
I've implemented pagination for the Alerts Module so their browsers
won't freeze trying to load a higher amount than usual files.

Note that the API side is sort of unusual due to the desire to place
less load on the database server. Each time a page is requested, all of
the user's file entries are retrieved and processed by the servlet
instead of letting the dataase deal with it. The processing seems to be
pretty quick on my computer even with thousands of files, so hopefully
it'll be fine.

Reducing the database load is a precaution due to the database being
resource starved recently.