Fix Content Collection link for course title.
The (view) link to the Content Collection next to the course titles
broke cause of the switch from displaying course name to course titles.

This is because the Content Collection uses the course name as part of
the path, not the course title.

This was fixed by transmitting the course name as well as the course
title in the json so we can use it as part of the Content Collection
path.