Remove unnecessary permissions.
Try to pare down permissions so we don't request too many of them.

The only one remaining that's a bit of a concern is read, write, execute
on all files. The execute may be removed if we don't need to execute
shell commands as part of the hostname discovery. Read and write is
required for the configuration file. The configuration file may be
disposed of if we just store all settings in the database. Which we may
end up doing since the configuration file may not be shared among
multiple nodes.