Convert /systemconfig/status to Restlet.
Convert the /systemconfig/status API calls from the previous standard
servlet into a Restlet enabled architecture.

The main motivation for moving to Restlet is to take advantage of the
better url handling that I had to do manually previously.

Note that the relative URLs in systemconfig.js were changed to include
/webapps/ubc-copyright-alerts-BBLEARN/systemconfig/ because of a Restlet
bug where urls of the form:
	/webapps/ubc-copyright-alerts-BBLEARN/systemconfig//status
wouldn't be recognized due to the double / before status. Regular
servlets collapse // down to one / but Restlet doesn't. This is a known
bug but unfortunately haven't been fixed (seems to be little interest in
fixing it).

The new SchedulerManager is a singleton class that creates and manages
all the Quartz scheduler associated machinery. It just needs to be
initialized and stopped by another servlet on startup for the scheduling
of index jobs to work.