Convert /systemconfig/schedule to Restlet.
This is the last of the system config API to be converted to Restlet.
With this one converted, the old SystemConfigAPI servlet is no longer
needed. It has been converted into the SchedulerStarter servlet, whose
purpose is to run application initialization code, which at this point
consists of starting the scheduler and storing the host into the
database.