Add API call for updating the index on demand.
The user can pass in an array of files in JSON format and it'll scan
those files for copyright tags, if the files have been tagged, it'll be
removed from the alerts database.

The JSON input should look like:
{"files":["/courses/AAAA000/inceptiontest/0000/0001.txt","/courses/AAAA000/inceptiontest/0000/0002.txt"]}

The original plan was to merge the metadata building block and the
copyright alerts building block. Looking at it technically, it turns out
that the effort required may be significant and we want the building
block to be out and available in half a month, so building an API for
communication between the two building blocks is more viable.