Rename methods in HostsTable for consistency.
Over time, I've added additional functionality so that the original
load(), save(), and delete() functions isn't clear in what they do.

They've been renamed to make it clear that you're performing operations
with respect to hosts.

Changed contains to hasHost since that's more in line with Java
conventions.

Also changed setRunStats to save since it's saving to database and not
just changing the internal state of the class.