Store schedule & metadata template config in db.
Just got confirmation that the B2 configuration file isn't shared across
nodes. So scheduling a task on one node would change the configuration
that one node but not for the other nodes.

I've switched SavedConfiguration to be stored in the Hosts table in the
config column. I've also switched SavedConfiguration to a singleton,
with better load/save control flow so that it's no longer being loaded
several times on startup or settings page view.

This is sorta hacky in that I've just put the plain text configuration
file that used to be stored on the file system into the database. This
should be sufficient for now unless the config grows more complex.

Note that we still need a mechanism for notifying the other nodes that
the configuration has changed.