Remove sync service, add hourly update.
Instead of notifying other nodes about config updates via the sync REST
service, have them refresh their configs every hour.

Added an hourly ConfigUpdateJob that does what the sync code used to do.

Note that a new trigger is always created with every scheduler update.
This new trigger doesn't inherit the state of the previous trigger,
which means that if the previous trigger is paused, the new trigger
isn't. This ended up confusing me while testing, since the scheduler
keeps pausing the trigger even though it previously paused it. What this
means is that there's a very small window where the trigger is created
and before it's paused that a job might execute even though it isn't
supposed to. The window is small though, so it probably isn't a
practical concern.