Skip non-course directories while indexing.
Apparently, it's possible for /courses to contain directories that are
not course related. There is now a CourseCache class that stores the
course id and title. The CourseCache will also check if a course exists
when it's asked to add a course. This should let us skip directories
that aren't course related.