Add progress indicator to systemconfig UI.
The progress indicator is just a count of the number of rows in the
queue and the number of rows in the files table.

The count is down by a SELECT COUNT(*) statement. My impression is that
this should be fine for medium sized tables. For large tables, a SELECT
COUNT(*) may not complete in reasonable time.