Share single instance of ConnectionManager.
Encountering connection exhausting issues in production. I think it
might be due to too many instances of ConnectionManager. I'm guessing
that ConnectionManager acts as a connection pool. This is only a guess
because documentation on this is non-existant. I've now made the entire
building block share a single instance of ConnectionManager, hopefully
it solves the problem.