No content system directory traversal with API.
Turns out that using the API to iterate through the content system
directories is a memory hog. Good news is that there seems be a table
that contains a flat version of the directory tree, so I'm now
generating the queue by reading from that table
(bblearn_cms_doc.xyf_urls).

This seems to have reduced the memory usage drastically, where it seems
to be only taking a constant amount of memory rather than polynomial.