Fix exception in exception cleanup.
In CSIndexJob, when an exception happens during indexing, an exception
cleanup method is called to make sure that execution status is properly
updated and that the remaining associated Quartz jobs are also cleaned
up. It turns out that an exception can happen during this cleanup,
hiding the original exception from logging. Rearranging the method so
that we log the error first before trying to cleanup now ensures that at
the very least, we have a log entry of the original exception.