Fix duplicate file alerts.
Some instructors may see a file show up more than once in alerts. This
can happen if they're teaching two courses that both use that file. When
grabbing the list of instructors that has permission to access that
file, it's a course based operation, so it'll list the instructor twice
since they teach both courses. Fix is to just change the list of users
to be stored as a set.