"Running" status now has 3 stages.
Previously, the status indicator in the configuration page would just
say "Running" and you wouldn't know if internally it's generating the
queue or actually processing the queue. This kind of information turns
out to be helpful in troubleshooting, so adding more detail to the
Running status was necessary.

I've now split the building block's running status into 3 stages:
* Queue Generation - Iterating the file system, looking for course
     files.
* Adding New Files - Going through the queue, adding in files that we
    haven't seen before.
* Updating Old Files - Goes through the files table, removing files that
    have been tagged. (To Be Implemented)

Updating the old files used to be done by the "Adding New Files" stage,
cause we drop the entire files table before generating the queue. It
looks like, due to the sheer amount of files in the system, that would
be counterproductive since it takes sooo long to generate the index. So
now I have to build a new stage for updating the old files instead of
dropping them all.