Add Queue Generation Resume support.
Turns out that there are so many files that queue generation may be
interrupted before it completes. So there needs to be a way to resume an
interrupted queue generation process.

Two fields were added to the HostsTable which stores the progress of the
queue generation when it's interrupted. See associated comments for what
they do. These fields are updated to the same value for all hosts.

The queue generation method was modified to save queue resume data every
BATCHSIZE files it encounter.

The SQL query used to retrieve files from where we left off may be a bit
slow due to the sheer amount of files we have to go through. It uses
ROW_NUMBER() to assign row numbers to each file. We're sorting the files
by file_id. file_id is an indexed field, which according to some
comments online, means that assigning row numbers *should* be very
efficient. Same with the sorting, because it's indexed, sorting
shouldn't take very much. We're using the location of the last processed
file as a kick off point for resuming queue generation. If that file was
deleted, then we just take a best guess by using the last row number
that we saw, which means some files may be missed. I'm assuming that
deletions don't happen that often, so missed files can just be picked
back up in the next cycle. Another assumption is that the file_id always
increases, it doesn't reuse old file_id, which *seems* to be the case.

Note that a previous attempt at the same SQL query used the "OFFSET"
mechanism to skip files, but it turns out that the version of Oracle
we're using is too old, so doesn't support "OFFSET" operations.