Add database reset functionality.
The database reset functionality just drops all data from this building
block's tables. It used to reside in a separate building block, but is
now integrated. The most complicated part is that clearing the database
also clears all of this building block's settings. This turned out to be
surprisingly problematic as many of the settings are cached. One thing
that I couldn't solve is that the UI requires a refresh in order to
retrieve the blank default settings. A message was put to ask the user
to refresh instead.