Conversion to long from int for resume data.
Resume data consists of an offset and a file id. Turns out that the INT
datatype we use in schema.xml defaults to 38 digit numbers in Oracle.
This is too large of a number for a Java int to hold, an SQL exception
is thrown when the conversion is impossible to make. Converting all the
relevant ints to longs should make this a non-issue.

Also refactored the resume data retrival methods in preparation for
implementing the update index resumable scanner.